adding logging and moving tests into integration folder
