+++ index.js
@@ -3,6 +3,7 @@ const express = require('express');
 const { ApolloServer } = require('apollo-server-express');
 const root = require('./src/root');
 const schema = require('./src/schema');
+const logger = require('./src/helpers/logger');
 
 const server = new ApolloServer({ typeDefs: schema, resolvers: root, playground: process.env.ENVIRONMENT === 'prod' ? false : true });
 const app = express();
@@ -23,8 +24,8 @@ require('dotenv').config({
     path: path.resolve(__dirname, `./${process.env.ENVIRONMENT}.env`)
 });
 
-app.listen({ port: 8081 }, () => {
-    console.log('Metacritic GraphQL API running on port 8081');
+app.listen({ port: process.env.PORT }, () => {
+    logger.info(`mcgqlapi running on port ${process.env.PORT}`);
 });
 
 module.exports = server;
