+++ src/helpers/logger.js
@@ -1,0 +1,12 @@
+const winston = require('winston');
+
+const logger = winston.createLogger({
+    level: 'info',
+    transports: [
+        new winston.transports.Console(),
+        new winston.transports.File({ filename: 'error.log', level: 'error' }),
+        new winston.transports.File({ filename: 'combined.log '})
+    ]
+});
+
+module.exports = logger;
